Steigung/Gefälle getter als int
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.sys.funclib.bitctrl@5966 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ src/de/bsvrz/sys/funclib/bitctrl/modell/verkehr/StrassenTeilSegment.java
@@ -114,7 +114,7 @@ public class StrassenTeilSegment extends StoerfallIndikator implements LinieXY {
 	 * 
 	 * @return Steigung (positiv) oder Gef&auml;lle (negativ) des Segments.
 	 */
-	public double getSteigungGefaelle() {
+	public int getSteigungGefaelle() {
 		leseKonfigDaten();
 		return steigungGefaelle;
 	}
